fix #27

+++ thus/slides.py
@@ -243,7 +243,13 @@ class Slides(GtkBaseBox):
     def reboot(self):
         """ Reboots the system, used when installation is finished """
         os.system("sync")
-        subprocess.call(["/usr/bin/systemctl", "reboot", "--force", "--no-wall"])
+        # OpenRC doesn't have systemctl
+        if os.path.exists("/usr/bin/systemctl"):
+            subprocess.call(["/usr/bin/systemctl", "reboot", "--force", "--no-wall"])
+        else:
+            subprocess.call(["/usr/bin/dbus-send", "--system", "--print-reply", 
+                             "--type=method_call", "--reply-timeout=2000", "--dest=org.freedesktop.ConsoleKit",
+                             "/org/freedesktop/ConsoleKit/Manager org.freedesktop.ConsoleKit.Manager.Restart"])
 
 # When testing, no _() is available
 try:
